requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.6.2
asgiref==3.8.1
async-property==0.2.2
async-timeout==4.0.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
bitarray==2.9.3
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.0
ckzg==2.0.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==43.0.0
cytoolz==1.0.0
daphne==4.1.2
deprecation==2.1.0
Django==5.1
django-celery==3.1.17
django-celery-beat==2.7.0
django-cors-headers==4.6.0
django-environ==0.11.2
django-timezone-field==7.1
djangorestframework==3.15.2
drf-spectacular==0.27.2
eth-account==0.13.4
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==2.1.0
eth-typing==5.0.0
eth-utils==5.0.0
eth_abi==5.1.0
exceptiongroup==1.2.2
flake8==7.1.1
frozenlist==1.4.1
gunicorn==23.0.0
h11==0.14.0
hexbytes==1.2.1
httpcore==1.0.6
httpx==0.27.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jwcrypto==1.5.6
kombu==5.4.2
mccabe==0.7.0
msgpack==1.1.0
multidict==6.1.0
packaging==24.1
parsimonious==0.10.0
pi==0.1.2
prompt_toolkit==3.0.50
propcache==0.2.0
psycopg2==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
pyflakes==3.2.0
pytz==2024.1
pyOpenSSL==25.0.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-keycloak==4.5.1
pyunormalize==16.0.0
PyYAML==6.0.2
redis==5.2.1
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rlp==4.0.1
rpds-py==0.20.0
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.1
tomli==2.2.1
toolz==1.0.0
Twisted==24.11.0
txaio==23.1.1
types-requests==2.32.0.20240914
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
web3==7.3.0
websockets==13.1
yarl==1.15.2
zope.interface==7.2
